refactor(desktop): move the context-menu policy into contextmenu.go
app.go held the Explorer "Send with Floe" policy in two places: the registry
key and the expected command near the top, and the four bound methods behind
the Settings switch further down, with the app lifecycle in between. They now
live together in contextmenu.go, beside the contextmenu_windows.go and
contextmenu_other.go pair that does the registry work. The file has no build
tag and no GOOS suffix on purpose: startup's self-heal calls into it behind a
runtime GOOS check, so it has to compile on Linux and macOS, and CI vets the
desktop module only on Windows. app.go drops the fmt import that only the
moved code used; packaged_windows.go's pointer to "the guards in app.go" and
DESKTOP.md's repo layout follow the code.

Pure move: go doc -all -u for the package is unchanged, so the *App method
set Wails reflects over is the same and frontend/wailsjs/go/main/App.d.ts is
byte-identical; go vet is clean on windows, GOOS=linux and GOOS=darwin; the
sorted go test -list output is identical; and git diff --color-moved shows
only package, import and header lines as new beside the two pointer edits.
